# swagger.yml
openapi: 3.0.2
info:
  title: API Ercole 2.0
  version: 1.0.0
  license:
    name: GNU General Public License v3.0 or later
    url: 'https://www.gnu.org/licenses/gpl-3.0-standalone.html'
  description: API used to talk to ercole services
  contact:
    name: Issues support
    url: 'https://github.com/ercole-io/ercole'
servers:
  - description: Api service on dev.ercole.io
    url: 'https://dev.ercole.io/api'
  - description: Data service on dev.ercole.io
    url: 'https://dev.ercole.io/data'
  - description: Alert service on dev.ercole.io
    url: 'https://dev.ercole.io/alert'
  - description: Repo service on dev.ercole.io
    url: 'https://dev.ercole.io/repo'
  - description: localhost address
    url: 'http://localhost:11111'
...
components:
  securitySchemes:
    basicAuthDataService:
      type: http
      scheme: basic
      description: access to data-service APIs
    basicAuthAlertService:
      type: http
      scheme: basic
      description: access to alert-service APIs
    basicAuthApiService:
      type: http
      scheme: basic
      description: access to api-service APIs
    tokenAuthApiService:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: access using the access token
  responses:
    error:
      description: Contains some informations about a error
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required:
              - Error
              - ErrorDescription
              - SourceFilename
              - LineNumber
            properties:
              Error:
                type: string
              ErrorDescription:
                type: string
              SourceFilename:
                type: string
              LineNumber:
                type: integer
          examples:
            Test1:
              value:
                Error: Invalid  status
                ErrorDescription: invalid status
                SourceFilename: /home/travis/go/src/github.com/ercole-io/ercole/api-service/controller/alerts_api.go
                LineNumber: 65
  schemas:
...
    PageMetadata:
      type: object
      required:
        - Empty
        - First
        - Last
        - Number
        - Size
        - TotalElements
        - TotalPages
      properties:
        Empty:
          type: boolean
        First:
          type: boolean
        Last:
          type: boolean
        Number:
          type: integer
        Size:
          type: integer
        TotalElements:
          type: integer
        TotalPages:
          type: integer
...
    PatchingFunction:
      type: object
      required:
        - Code
        - Vars
      properties:
        ID:
          type: string
          nullable: true
        Hostname:
          type: string
          nullable: true
        CreatedAt:
          type: string
          nullable: true
        Code:
          type: string
        Vars:
          type: object
...
    LicenseModifier:
      type: object
      required:
        - _id
        - Hostname
        - DatabaseName
        - LicenseName
        - NewValue
      properties:
        _id: 
          type: string
        Hostname:
          type: string
        DatabaseName:
          type: string
        LicenseName:
          type: string
        NewValue:
          type: number
  parameters:
    search:
      in: query
      name: search
      required: false
      description: filter the result using search terms
      allowEmptyValue: true
      example: foobar
      schema:
        type: string
        readOnly: true
        example: foobar
    sort-by:
      in: query
      name: sort-by
      required: false
      description: sort the result by a field
      allowEmptyValue: true
      example: foobar
      schema:
        type: string
        readOnly: true
        example: CreatedAt
    sort-desc:
      schema:
        type: boolean
      in: query
      name: sort-desc
      description: |
        Sort the field descendently instead of ascendently
      allowEmptyValue: true
    page:
      schema:
        type: integer
      in: query
      name: page
      description: Number of the page
      allowEmptyValue: true
    size:
      schema:
        type: integer
      in: query
      name: size
      description: Size of the page
      allowEmptyValue: true
...
paths:
...
  '/hosts/{hostname}/patching-function':
    parameters:
      - in: path
        name: hostname
        schema:
          type: string
        required: true
        description: hostname of the requested host
    get:
      tags:
        - api-service
        - developer-user
        - read
      operationId: GetPatchingFunction
      summary: Get a patching function of a host
      description: Get a patching function of a host
      responses:
        '200':
          description: Patching function of the host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatchingFunction'
        '401':
          $ref: '#/components/responses/error'
        '404':
          description: 'The host or patching function don"t exist'
          $ref: '#/components/responses/error'
        '422':
          $ref: '#/components/responses/error'
        '500':
          $ref: '#/components/responses/error'
    put:
      tags:
        - api-service
        - developer-user
        - write
      operationId: SetPatchingFunction
      summary: Set a patching function of a host
      description: Set the patching function of a host
      responses:
        '200':
          description: Patching function of the host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatchingFunction'
        '401':
          $ref: '#/components/responses/error'
        '404':
          description: 'The host don"t exist'
          $ref: '#/components/responses/error'
        '422':
          $ref: '#/components/responses/error'
        '500':
          $ref: '#/components/responses/error'
  '/hosts/{hostname}/technologies/oracle/databases/{dbname}/tags':
    parameters:
      - in: path
        name: hostname
        schema:
          type: string
        required: true
        description: hostname of the requested host
      - in: path
        name: dbname
        schema:
          type: string
        required: true
        description: name of the modifying database
    post:
      tags:
        - api-service
        - fe-user
        - write
      operationId: AddTagToOracleDatabase
      summary: Add a tag to a database of a host
      description: Add a tag to a database of a host
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
      responses:
        '200':
          description: The tag was successfully attached to the db
          content:
            application/json: {}
        '401':
          $ref: '#/components/responses/error'
        '422':
          $ref: '#/components/responses/error'
        '500':
          $ref: '#/components/responses/error'
  '/hosts/{hostname}/technologies/oracle/databases/{dbname}/tags/{tagname}':
    parameters:
      - in: path
        name: hostname
        schema:
          type: string
        required: true
        description: hostname of the requested host
      - in: path
        name: dbname
        schema:
          type: string
        required: true
        description: name of the modifying database
      - in: path
        name: tagname
        schema:
          type: string
        required: true
        description: name of the tag
    delete:
      tags:
        - api-service
        - fe-user
        - write
      operationId: DeleteTagOfOracleDatabase
      summary: Delete a tag of a database of a host
      description: Add a tag of a database of a host
      responses:
        '200':
          description: The tag was successfully deleted from the db
          content:
            application/json: {}
        '401':
          $ref: '#/components/responses/error'
        '422':
          $ref: '#/components/responses/error'
        '500':
          $ref: '#/components/responses/error'
  '/hosts/{hostname}/technologies/oracle/databases/{dbname}/licenses/{licenseName}':
    parameters:
      - in: path
        name: hostname
        schema:
          type: string
        required: true
        description: hostname of the requested host
      - in: path
        name: dbname
        schema:
          type: string
        required: true
        description: name of the modifying database
      - in: path
        name: licenseName
        schema:
          type: string
        required: true
        description: name of the license
    put:
      tags:
        - api-service
        - fe-user
        - write
      operationId: SetOracleDatabaseLicenseModifier
      summary: Change the value of a license of a database
      description: Change the value of a license of a database
      requestBody:
        content:
          text/plain:
            schema:
              type: number
      responses:
        '200':
          description: The value of the the license was changed successfully
          content:
            application/json: {}
        '401':
          $ref: '#/components/responses/error'
        '422':
          $ref: '#/components/responses/error'
        '500':
          $ref: '#/components/responses/error'
  '/hosts/{hostname}/technologies/oracle/databases/{dbname}/license-modifiers/{licenseName}':
    parameters:
      - in: path
        name: hostname
        schema:
          type: string
        required: true
        description: hostname of the requested host
      - in: path
        name: dbname
        schema:
          type: string
        required: true
        description: name of the modifying database
      - in: path
        name: licenseName
        schema:
          type: string
        required: true
        description: name of the license
    delete:
      tags:
        - api-service
        - fe-user
        - write
      operationId: DeleteOracleDatabaseLicenseModifier
      summary: Delete a license modifier of a database
      description: Delete a license modifier of a database
      responses:
        '200':
          description: The license modifier was deleted successfully
          content:
            application/json: {}
        '401':
          $ref: '#/components/responses/error'
        '422':
          $ref: '#/components/responses/error'
        '500':
          $ref: '#/components/responses/error'
...
  /hosts/technologies/oracle/databases/license-modifiers:
    get:
      tags:
        - api-service
        - fe-user
        - read
      operationId: SearchOracleDatabaseLicenseModifiers
      summary: Search a list of license modifiers
      description: Get a list of license modifiers using various search terms and various params.
      parameters:
        - $ref: '#/components/parameters/search'
        - $ref: '#/components/parameters/sort-by'
        - $ref: '#/components/parameters/sort-desc'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Result of the search
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    description: paged
                    properties:
                      Content:
                        type: array
                        items:
                          $ref: '#/components/schemas/LicenseModifier'
                      Metadata:
                        $ref: '#/components/schemas/PageMetadata'
                    required:
                      - Content
                      - Metadata
                  - type: array
                    description: non paged
                    items:
                      $ref: '#/components/schemas/LicenseModifier'
        '401':
          $ref: '#/components/responses/error'
        '422':
          $ref: '#/components/responses/error'
        '500':
          $ref: '#/components/responses/error'
...
security:
  - basicAuthApiService: []
  - tokenAuthApiService: []